+++ src/route/Top.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Top = () => {
+  return <div>Top</div>;
+};
+
+export default Top;
