+++ src/three_befo_change_poti.tsx
@@ -3,12 +3,9 @@ import { createRoot } from "react-dom/client";
 import "./index.css";
 import App from "./route/Home.tsx";
 import React from "react";
-import { BrowserRouter } from "react-router";
 
-createRoot(document.getElementById("root")!).render(
+createRoot(document.getElementById("main")!).render(
   <StrictMode>
-    <BrowserRouter>
-      <App />
-    </BrowserRouter>
+    <App />
   </StrictMode>
 );
